refactor: Extract getPlotsTotalDistance from Interpolator to Application Layer's Service Directory
